+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.0
+## 2016-09-27
+* Added a `purgeEverything` function to CloudflarePurger.
+* Tweak CloudflarePurgerExtension to detect navigation changes and trigger a full purge.
+
+## 2016-09-26
+* Add logic to Purge draft of pages.
+
 # 0.0.1
 ## 2016-09-05
 * Fix a bug to make sure URL with and wihtout a trailing slash are purge.